Menambahkan file upload avatar

+++ pemilik/lihat-user.php
@@ -1,19 +1,21 @@
-<?php 
+<?php
 include '../koneksi.php';
 
 ?>
 <!DOCTYPE html>
 <html>
+
 <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">	
+	<meta charset="utf-8">
+	<meta name="viewport" content="width=device-width, initial-scale=1.0">
 	<title>Lihat User</title>
 	<link rel="stylesheet" type="text/css" href="../style.css">
-    <link rel="stylesheet" type="text/css" href="../font-awesome/css/font-awesome.min.css">
+	<link rel="stylesheet" type="text/css" href="../font-awesome/css/font-awesome.min.css">
 </head>
+
 <body>
 	<main class="content">
-		<?php include'menu-content.php';?>
+		<?php include 'menu-content.php'; ?>
 		<section class="views">
 			<div class="search">
 				<input class="search-data" placeholder="Cari"></input>
@@ -21,27 +23,27 @@ include '../koneksi.php';
 			<div class="view-table">
 				<div class="welcome">
 					<h1>Lihat User</h1>
-					<hr/>
+					<hr />
 				</div>
 				<div class="table">
 					<table>
 						<thead>
-							<tr>	
-								<th>No</th>			
-								<th>ID USER</th>					
-								<th>Nama</th>					
-								<th>Password</th>					
-								<th>Hak Akses</th>					
-								<th>Tanggal Dibuat</th>					
+							<tr>
+								<th>No</th>
+								<th>ID USER</th>
+								<th>Nama</th>
+								<th>Password</th>
+								<th>Hak Akses</th>
+								<th>Tanggal Dibuat</th>
 								<th>Manager</th>
-								<th>Update - Delete</th>										
+								<th>Update - Delete</th>
 							</tr>
 						</thead>
 						<tbody>
-						<?php
-                            $sql = mysqli_query($koneksi, "SELECT * FROM user ");
-                            $no = 1;
-                            while ($r = mysqli_fetch_array($sql)) {
+							<?php
+							$sql = mysqli_query($koneksi, "SELECT * FROM user ");
+							$no = 1;
+							while ($r = mysqli_fetch_array($sql)) {
 								echo "<tr>
 										<td>$no</td>	
 										<td>$r[id_user]</td>				
@@ -52,19 +54,19 @@ include '../koneksi.php';
 										<td>$r[manager]</td>								
 										<td>
 											<a href='aksi_edit_user.php?id_user=$r[id_user]'><button>Update</button></a>										
-											<a href='aksi_hapus_user.php?id_user=$r[id_user]'><button>Delete</button></a>
-										</td>";	
+											<a href='aksi_hapus_user.php?id_user=$r[id_user]' onclick='return confirm('Apakah anda yakin ingin menghapus?')'><button>Delete</button></a>
+										</td>";
 								$no++;
-								}
-							
-							?>														
+							}
+
+							?>
 						</tbody>
 					</table>
 				</div>
 			</div>
 		</section>
 	</main>
-	
+
 	<footer>
 		<div class="wrapper">
 			<div class="copyright">
@@ -72,6 +74,7 @@ include '../koneksi.php';
 			</div>
 		</div>
 	</footer>
-		<script src="../bootstrap.bundle.min.js"></script>
+	<script src="../bootstrap.bundle.min.js"></script>
 </body>
+
 </html>
